created local module for auxiliary functions
