+++ 1-node-farm/starter/index.js
@@ -3,6 +3,8 @@ const http = require('http');
 const { json } = require('stream/consumers');
 const url = require('url');
 
+const replaceTemplate = require('./modules/replaceTemplate')
+
 //*****************************************************
 //FILES
 
@@ -34,22 +36,6 @@ console.log("File written!");
 
 //******************************************* 
 // SERVER
-
-const replaceTemplate = (template, product) => {
-    let output = template.replace(/{%PRODUCTNAME%}/g, product.productName)
-    output = output.replace(/{%IMAGE%}/g, product.image)
-    output = output.replace(/{%PRICE%}/g, product.price)
-    output = output.replace(/{%FROM%}/g, product.from)
-    output = output.replace(/{%NUTRIENTS%}/g, product.nutrients)
-    output = output.replace(/{%QUANTITY%}/g, product.quantity)
-    output = output.replace(/{%DESCRIPTION%}/g, product.description)
-    output = output.replace(/{%ID%}/g, product.id)
-
-    if(!product.organic){
-        output = output.replace(/{%NOT_ORGANIC%}/g, 'not-organic')
-    }
-    return output;
-}
 
 //Top level code which is executed only once, hence we can use the synchronous version here
 const templateOverview = fs.readFileSync("./templates/template-overview.html", 'utf-8')
